move adapters to an own package
