+++ app/src/main/java/es/wolfi/app/passman/adapters/CustomFieldEditAdapter.java
@@ -18,7 +18,7 @@
  * You should have received a copy of the GNU Affero General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package es.wolfi.app.passman;
+package es.wolfi.app.passman.adapters;
 
 import android.app.ProgressDialog;
 import android.content.Context;
@@ -49,6 +49,7 @@ import java.util.Timer;
 import java.util.TimerTask;
 
 import es.wolfi.app.ResponseHandlers.CustomFieldFileDeleteResponseHandler;
+import es.wolfi.app.passman.R;
 import es.wolfi.passman.API.Credential;
 import es.wolfi.passman.API.CustomField;
 import es.wolfi.passman.API.File;
