+++ app/src/main/java/es/wolfi/app/passman/adapters/FileEditAdapter.java
@@ -18,7 +18,7 @@
  * You should have received a copy of the GNU Affero General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package es.wolfi.app.passman;
+package es.wolfi.app.passman.adapters;
 
 import android.app.ProgressDialog;
 import android.content.Context;
@@ -38,6 +38,7 @@ import org.json.JSONException;
 import java.util.List;
 
 import es.wolfi.app.ResponseHandlers.FileDeleteResponseHandler;
+import es.wolfi.app.passman.R;
 import es.wolfi.passman.API.Credential;
 import es.wolfi.passman.API.File;
 import es.wolfi.utils.FileUtils;
