+++ app/src/main/java/es/wolfi/app/passman/adapters/VaultViewAdapter.java
@@ -20,7 +20,7 @@
  *
  */
 
-package es.wolfi.app.passman;
+package es.wolfi.app.passman.adapters;
 
 import androidx.recyclerview.widget.RecyclerView;
 import android.util.Log;
@@ -34,6 +34,7 @@ import java.util.List;
 
 import butterknife.BindView;
 import butterknife.ButterKnife;
+import es.wolfi.app.passman.R;
 import es.wolfi.app.passman.VaultFragment.OnListFragmentInteractionListener;
 import es.wolfi.passman.API.Vault;
 import es.wolfi.utils.ColorUtils;
