+++ app/src/main/java/es/wolfi/utils/FilterListAsyncTask.java
@@ -29,13 +29,11 @@ import androidx.recyclerview.widget.RecyclerView;
 import java.util.ArrayList;
 
 import es.wolfi.app.passman.CredentialItemFragment;
-import es.wolfi.app.passman.CredentialViewAdapter;
+import es.wolfi.app.passman.adapters.CredentialViewAdapter;
 import es.wolfi.app.passman.VaultFragment;
-import es.wolfi.app.passman.VaultViewAdapter;
+import es.wolfi.app.passman.adapters.VaultViewAdapter;
 import es.wolfi.passman.API.Credential;
 import es.wolfi.passman.API.Vault;
-import es.wolfi.utils.Filterable;
-import es.wolfi.utils.ListUtils;
 
 public class FilterListAsyncTask <T extends Filterable> extends AsyncTask<ArrayList<T>, Integer, ArrayList<T>>{
